otel-demo-app/requirements.txt
fastapi
uvicorn

opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation
psutil
